Reformatted, moved to community package and cleaned up
