+++ src/main/java/org/infinispan/visualizer/poller/infinispan/JdgJmxCacheNamesPollerManager.java
@@ -1,0 +1,56 @@
+package org.infinispan.visualizer.poller.infinispan;
+
+import java.net.InetSocketAddress;
+import java.net.MalformedURLException;
+import java.net.SocketAddress;
+import java.util.Map;
+
+import javax.management.remote.JMXServiceURL;
+
+import org.infinispan.visualizer.internal.VisualizerRemoteCacheManager;
+import org.infinispan.visualizer.poller.jmx.JmxCacheNamesPollerManager;
+import org.infinispan.visualizer.poller.jmx.JmxPoller;
+
+/*
+* JBoss, Home of Professional Open Source
+* Copyright 2011 Red Hat Inc. and/or its affiliates and other
+* contributors as indicated by the @author tags. All rights reserved.
+* See the copyright.txt in the distribution for a full listing of
+* individual contributors.
+*
+* This is free software; you can redistribute it and/or modify it
+* under the terms of the GNU Lesser General Public License as
+* published by the Free Software Foundation; either version 2.1 of
+* the License, or (at your option) any later version.
+*
+* This software is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+* Lesser General Public License for more details.
+*
+* You should have received a copy of the GNU Lesser General Public
+* License along with this software; if not, write to the Free
+* Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+* 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+*/
+public class JdgJmxCacheNamesPollerManager extends JmxCacheNamesPollerManager {
+
+   public JdgJmxCacheNamesPollerManager(VisualizerRemoteCacheManager cacheManager) {
+      super(cacheManager);
+   }
+
+   @Override
+   protected JMXServiceURL generateServiceURL(SocketAddress address)
+         throws MalformedURLException {
+      InetSocketAddress isa = (InetSocketAddress) address;
+      String host = isa.getAddress().getHostAddress();
+      int port = isa.getPort() - getJmxHotrodPortOffset();
+      return new JMXServiceURL("service:jmx:http-remoting-jmx://" + host + ":" + port);
+   }
+
+   @Override
+   protected JmxPoller<String[]> createPoller(JMXServiceURL url,
+                                              Map<String, Object> env) {
+      return new JdgJmxCacheNamesPoller(url, env);
+   }
+}
