+++ src/main/java/org/infinispan/visualizer/poller/jmx/JmxCacheEntriesPollerManager.java
@@ -1,0 +1,66 @@
+/*
+* JBoss, Home of Professional Open Source
+* Copyright 2011 Red Hat Inc. and/or its affiliates and other
+* contributors as indicated by the @author tags. All rights reserved.
+* See the copyright.txt in the distribution for a full listing of
+* individual contributors.
+*
+* This is free software; you can redistribute it and/or modify it
+* under the terms of the GNU Lesser General Public License as
+* published by the Free Software Foundation; either version 2.1 of
+* the License, or (at your option) any later version.
+*
+* This software is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+* Lesser General Public License for more details.
+*
+* You should have received a copy of the GNU Lesser General Public
+* License along with this software; if not, write to the Free
+* Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+* 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+*/
+package org.infinispan.visualizer.poller.jmx;
+
+import java.net.SocketAddress;
+
+import org.infinispan.visualizer.internal.VisualizerRemoteCacheManager;
+import org.infinispan.visualizer.rest.NodeInfo;
+
+/**
+ * @author <a href="mailto:rtsang@redhat.com">Ray Tsang</a>
+ */
+public abstract class JmxCacheEntriesPollerManager extends JmxPollerManager<NodeInfo, Integer, JmxCacheEntriesPoller> {
+   private volatile int colorIndex = 0;
+   private boolean multiColor = false;
+
+   private String cacheName;
+
+   public JmxCacheEntriesPollerManager(
+         VisualizerRemoteCacheManager cacheManager) {
+      super(cacheManager);
+   }
+
+   @Override
+   protected NodeInfo createNewInfo(String id, SocketAddress addr) {
+      if (multiColor)
+         return new NodeInfo(id, addr.toString(), colorIndex++);
+      else
+         return new NodeInfo(id, addr.toString(), colorIndex);
+   }
+
+   public String getCacheName() {
+      return cacheName;
+   }
+
+   public void setCacheName(String cacheName) {
+      this.cacheName = cacheName;
+   }
+
+   public void setMultiColor(Boolean multiColor, Integer colorIndex) {
+      this.multiColor = multiColor;
+      if (!this.multiColor) {
+         this.colorIndex = colorIndex;
+      }
+   }
+}
