+++ src/main/java/org/infinispan/visualizer/poller/jmx/JmxCacheNamesPollerManager.java
@@ -20,51 +20,32 @@
 * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
 * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
 */
+package org.infinispan.visualizer.poller.jmx;
 
-/* 
-    Document   : viz.css
-    Created on : Mar 29, 2011, 3:22:43 PM
-    Author     : Andrew Sacamano<andrew.sacamano@amentra.com>
-    Description:
-    A stylesheet for the Infinispan visualizer.
-*/
+import java.net.SocketAddress;
 
-body {
-    color: #fff;
-    background: #323232 url(grey_tile.png) repeat-x;
+import org.infinispan.visualizer.internal.VisualizerRemoteCacheManager;
+import org.infinispan.visualizer.poller.CacheNamesPollerThread;
+import org.infinispan.visualizer.rest.CacheNameInfo;
+
+/**
+ * @author <a href="mailto:rtsang@redhat.com">Ray Tsang</a>
+ */
+public abstract class JmxCacheNamesPollerManager extends JmxPollerManager<CacheNameInfo, String[], JmxPoller<String[]>> {
+
+   public JmxCacheNamesPollerManager(VisualizerRemoteCacheManager cacheManager) {
+      super(cacheManager);
+   }
+
+   @Override
+   protected CacheNamesPollerThread createPollerThread(SocketAddress address,
+                                                       CacheNameInfo info) throws Exception {
+      return new CacheNamesPollerThread(createPoller(address), info);
+   }
+
+   @Override
+   protected CacheNameInfo createNewInfo(String id, SocketAddress addr) {
+      return new CacheNameInfo();
+   }
+
 }
-
-#stage {
-    width: 100%;
-    height: 100%;
-    position: relative;
-    margin: 0 auto;
-    padding: 0;
-    z-index: -1;
-}
-
-.node {
-    width: 150px;
-    height: 150px;
-    display: none;
-    position: absolute;
-    top: 0;
-    left: 0;
-    text-align: center;
-}
-
-.nodetitle {
-    /* display: none; */
-    z-index: -1;
-}
-
-.nodecanvas {
-    width: 150px;
-    height: 150px;
-}
-
-#controls {
-    float: left;
-    z-index: 100;
-}
-
