+++ src/main/webapp/WEB-INF/beans.xml
@@ -3,7 +3,7 @@
 <!-- We're declaring the schema to save you time if you do have to configure 
    this in the future -->
 <beans xmlns="http://java.sun.com/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-   xsi:schemaLocation="
+       xsi:schemaLocation="
         http://java.sun.com/xml/ns/javaee 
         http://java.sun.com/xml/ns/javaee/beans_1_0.xsd">
 </beans>
